@@ -6,10 +6,8 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
-    "p5": "^1.1.9",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-p5-wrapper": "^2.3.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^5.0.1",
     "react-transition-group": "^4.4.1"
